Support IMAP-to-IMAP copy
Patch received via e-mail from Zane Selvans:
"""
Hi there,

Thanks so much for posting your imap2maildir Python script!  I'm using it to
migrate tens of gigabytes of my friends-and-family mail archives.  I needed
to be able to reliably send the downloaded mail back up to another IMAP
server, in a way that drag-and-drop within Apple Mail and Thunderbird just
couldn't do, so I added the option of specifying a second IMAP folder as the
destination, in addition to the Maildir and mbox type destinations.

When run the first time, it adds the hashes of all the messages on the
destination IMAP server to the DB so that any messages that *did*
successfully get transferred in the user's failed drag-and-drop attempts at
migration don't get duplicated.

I've attached the updated script if you feel like distributing it.  It isn't
extensively tested, but I did update the usage/help messages.

Cheers,
Zane
"""

TODO: Extensively test it :-)  -rt